super-agent/requirements.txt
# Super-Agent V2.1 — Phase 1 (Foundation)
# Phase 3+: add langgraph + langchain-core (see requirements-orchestrator.txt)
pydantic>=2.5.0,<3
httpx>=0.27.0,<1
pyyaml>=6.0.1,<7
openai>=1.55.0,<2
